backend/requirements.txt
fastapi 
uvicorn[standard] 
pandas 
scikit-learn 
pymongo
python-dotenv
transformers
torch